Added react-leaflet package, resolved package errors
I had to resolve some package dependencies errors. material-ui does not yet work for React 17, but the newest
version of react-leaflet required it.

I ended up downgrading react (and react-dom) to 16.14 and downgrading react-leaflet to 2.8.0 for everything to
be compatible with material-ui.

NOTE: When material-ui releases v5, upgrade everything back to the latest version.